requirements.txt
aio-pika==6.4.1
aiormq==3.2.0
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.9.11
chardet==3.0.4
Django==3.0.1
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
pika==1.1.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7